pyproject.toml
[project]
name = "pdf-highlight-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "pdf2image>=1.17.0",
    "pillow>=11.3.0",
    "pymupdf>=1.26.4",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "mypy>=1.17.1",
    "pytest>=8.4.1",
    "ruff>=0.12.11",
]
